fix(parsing): Move all parsing of single f64s to string parser
Single f64s do not get error-corrected, causing them to simply be wrong (at least at the precision
we want) in some cases. I'm not convinced that parsing them as strings is the answer, but for the
moment it'll do. This will be revisited when the library is functionally complete.